Fix build

+++ csrc/ir/iostream.h
@@ -9,9 +9,7 @@
 
 #include <iosfwd>
 
-#include "dispatch.h"
-#include "exceptions.h"
-#include "visibility.h"
+#include "base.h"
 
 namespace nvfuser {
 
@@ -20,7 +18,7 @@ void checkInlineable(const Expr* expr);
 static constexpr char const* kTab = "  ";
 
 // Indent the generated code
-inline std::ostream& indent(std::ostream& os, int indent_size) {
+inline std::ostream& indent(std::ostream& os, int64_t indent_size) {
   for (const auto _ : arange(indent_size)) {
     (void)_; // Suppress unused variable warning
     os << "  ";
